refactor (PurchaseAmountDto): 구입 금액 검증 기능 이동 PurchaseAmount -> PurchaseAmountDto
