+++ src/test/java/lotto/domain/PurchaseAmountTest.java
@@ -2,6 +2,7 @@ package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.domain.dto.PurchaseAmountDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -9,8 +10,8 @@ import org.junit.jupiter.api.Test;
 class PurchaseAmountTest {
     @Nested
     @DisplayName("validatePurchaseAmount 메소드 test")
-    class ValidatePurchaseAmount {
-        @DisplayName("구입 금액이 " + PurchaseAmount.PURCHASE_AMOUNT_UNIT + " 단위면 검증 통과")
+    class ValidatePurchaseAmountDto {
+        @DisplayName("구입 금액이 " + PurchaseAmountDto.PURCHASE_AMOUNT_UNIT + " 단위면 검증 통과")
         @Test
         void Amount_is_divisible_by_the_purchase_amount_unit() {
             // given
@@ -20,12 +21,12 @@ class PurchaseAmountTest {
 
             // when
             // then
-            PurchaseAmount purchaseAmount1 = new PurchaseAmount(amount1);
-            PurchaseAmount purchaseAmount2 = new PurchaseAmount(amount2);
-            PurchaseAmount purchaseAmount3 = new PurchaseAmount(amount3);
+            PurchaseAmountDto purchaseAmount1 = new PurchaseAmountDto(amount1);
+            PurchaseAmountDto purchaseAmount2 = new PurchaseAmountDto(amount2);
+            PurchaseAmountDto purchaseAmount3 = new PurchaseAmountDto(amount3);
         }
 
-        @DisplayName("구입 금액이 " + PurchaseAmount.PURCHASE_AMOUNT_UNIT + " 단위가 아니면 예외 발생")
+        @DisplayName("구입 금액이 " + PurchaseAmountDto.PURCHASE_AMOUNT_UNIT + " 단위가 아니면 예외 발생")
         @Test
         void test() {
             // given
@@ -34,12 +35,10 @@ class PurchaseAmountTest {
 
             // when
             // then
-            assertThatThrownBy(() -> new PurchaseAmount(amount1))
-                    .isInstanceOf(IllegalArgumentException.class)
-                    .hasMessage(PurchaseAmount.PURCHASE_AMOUNT_NOT_DIVISIBLE_MESSAGE);
-            assertThatThrownBy(() -> new PurchaseAmount(amount2))
-                    .isInstanceOf(IllegalArgumentException.class)
-                    .hasMessage(PurchaseAmount.PURCHASE_AMOUNT_NOT_DIVISIBLE_MESSAGE);
+            assertThatThrownBy(() -> new PurchaseAmountDto(amount1)).isInstanceOf(IllegalArgumentException.class)
+                    .hasMessage(PurchaseAmountDto.PURCHASE_AMOUNT_NOT_DIVISIBLE_MESSAGE);
+            assertThatThrownBy(() -> new PurchaseAmountDto(amount2)).isInstanceOf(IllegalArgumentException.class)
+                    .hasMessage(PurchaseAmountDto.PURCHASE_AMOUNT_NOT_DIVISIBLE_MESSAGE);
         }
     }
 }
