+++ src/test/java/lotto/service/LottoMachineTest.java
@@ -2,7 +2,6 @@ package lotto.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.domain.PurchaseAmount;
 import lotto.domain.dto.LottosDto;
 import lotto.domain.dto.PurchaseAmountDto;
 import org.junit.jupiter.api.DisplayName;
@@ -13,7 +12,7 @@ class LottoMachineTest {
     @Nested
     @DisplayName("issuedLottos 메소드 test")
     class IssuedLottos {
-        @DisplayName("발행된 로또 수는 구입 금액을 " + PurchaseAmount.PURCHASE_AMOUNT_UNIT + "으로 나눈 값과 같다")
+        @DisplayName("발행된 로또 수는 구입 금액을 " + PurchaseAmountDto.PURCHASE_AMOUNT_UNIT + "으로 나눈 값과 같다")
         @Test
         void Issued_lotto_count_is_equal_to_purchase_amount() {
             // given
@@ -25,7 +24,7 @@ class LottoMachineTest {
 
             // then
             assertThat(lottosDto.lottos().size()).isEqualTo(
-                    purchaseAmountDto.amount() / PurchaseAmount.PURCHASE_AMOUNT_UNIT);
+                    purchaseAmountDto.amount() / PurchaseAmountDto.PURCHASE_AMOUNT_UNIT);
         }
     }
 }
